# pkg/httphandler/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "App API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "Hello world.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/items": {
            "get": {
                "tags": [
                    "items"
                ],
                "summary": "Get items.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.ItemsListResponse"
                            }
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "tags": [
                    "items"
                ],
                "summary": "Get item by id.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ItemModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.ItemModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "app.ItemsListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.ItemModel"
                    }
                }
            }
        },
        "httphandler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}